Move analysis wrapper to separate file.
