+++ iree/compiler/Dialect/VM/Conversion/VMToEmitC/ConvertVMToEmitC.h
@@ -7,62 +7,13 @@
 #ifndef IREE_COMPILER_DIALECT_VM_CONVERSION_VMTOEMITC_CONVERTVMTOEMITC_H_
 #define IREE_COMPILER_DIALECT_VM_CONVERSION_VMTOEMITC_CONVERTVMTOEMITC_H_
 
-#include "iree/compiler/Dialect/VM/Analysis/RegisterAllocation.h"
-#include "iree/compiler/Dialect/VM/Analysis/ValueLiveness.h"
 #include "iree/compiler/Dialect/VM/Conversion/VMToEmitC/EmitCTypeConverter.h"
+#include "iree/compiler/Dialect/VM/Conversion/VMToEmitC/VMAnalysis.h"
 #include "iree/compiler/Dialect/VM/IR/VMOps.h"
-#include "mlir/Dialect/EmitC/IR/EmitC.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
 namespace iree_compiler {
-
-struct VMAnalysis {
- public:
-  VMAnalysis(RegisterAllocation &&registerAllocation,
-             ValueLiveness &&valueLiveness)
-      : registerAllocation(std::move(registerAllocation)),
-        valueLiveness(std::move(valueLiveness)) {}
-
-  VMAnalysis(VMAnalysis &&) = default;
-  VMAnalysis &operator=(VMAnalysis &&) = default;
-  VMAnalysis(const VMAnalysis &) = delete;
-  VMAnalysis &operator=(const VMAnalysis &) = delete;
-
-  int getNumRefRegisters() {
-    return registerAllocation.getMaxRefRegisterOrdinal() + 1;
-  }
-
-  uint16_t getRefRegisterOrdinal(Value ref) {
-    assert(ref.getType().isa<IREE::VM::RefType>());
-    return registerAllocation.mapToRegister(ref).ordinal();
-  }
-
-  bool isLastValueUse(Value ref, Operation *op) {
-    assert(ref.getType().isa<IREE::VM::RefType>());
-    return valueLiveness.isLastValueUse(ref, op);
-  }
-
-  void cacheLocalRef(int64_t ordinal, emitc::ApplyOp &applyOp) {
-    assert(!refs.count(ordinal));
-    refs[ordinal] = applyOp.getOperation();
-  }
-
-  emitc::ApplyOp lookupLocalRef(int64_t ordinal) {
-    assert(refs.count(ordinal));
-    Operation *op = refs[ordinal];
-    return cast<emitc::ApplyOp>(op);
-  }
-
-  DenseMap<int64_t, Operation *> &localRefs() { return refs; }
-
- private:
-  RegisterAllocation registerAllocation;
-  ValueLiveness valueLiveness;
-  DenseMap<int64_t, Operation *> refs;
-};
-
-using VMAnalysisCache = DenseMap<Operation *, VMAnalysis>;
 
 void populateVMToEmitCPatterns(MLIRContext *context,
                                IREE::VM::EmitCTypeConverter &typeConverter,
